requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
xgboost>=2.0.0
streamlit>=1.36.0
plotly>=5.22.0
joblib>=1.4.0
requests>=2.32.0
pyyaml>=6.0.1
python-dateutil>=2.9.0
